+++ src/main/java/br/com/infocedro/medflow/intake/IntakeRepository.java
@@ -1,5 +1,11 @@
 package br.com.infocedro.medflow.intake;
 
-public class IntakeRepository {
-    
+import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface IntakeRepository extends JpaRepository<Intake, Long> {
+
+    List<Intake> findAllByActiveTrueOrderByScheduledAtAsc();
+
+    List<Intake> findAllByPrescriptionIdAndActiveTrueOrderByScheduledAtAsc(Long prescriptionId);
 }