+++ src/main/java/br/com/infocedro/medflow/intake/IntakeStatus.java
@@ -1,0 +1,7 @@
+package br.com.infocedro.medflow.intake;
+
+public enum IntakeStatus {
+    PENDING,
+    TAKEN,
+    SKIPPED
+}